Force logs whitelisted logs
