+++ src/Azure.Functions.Cli/Actions/HostActions/StartHostAction.cs
@@ -6,27 +6,19 @@ using System.Linq;
 using System.Net;
 using System.Security.Cryptography.X509Certificates;
 using System.Threading.Tasks;
-using Azure.Functions.Cli.Actions.HostActions.WebHost.Security;
 using Azure.Functions.Cli.Common;
 using Azure.Functions.Cli.Diagnostics;
-using Azure.Functions.Cli.ExtensionBundle;
 using Azure.Functions.Cli.Extensions;
 using Azure.Functions.Cli.Helpers;
 using Azure.Functions.Cli.Interfaces;
 using Azure.Functions.Cli.NativeMethods;
 using Colors.Net;
 using Fclp;
-using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Azure.WebJobs.Extensions.Http;
 using Microsoft.Azure.WebJobs.Script;
 using Microsoft.Azure.WebJobs.Script.Description;
 using Microsoft.Azure.WebJobs.Script.WebHost;
-using Microsoft.Azure.WebJobs.Script.WebHost.Authentication;
-using Microsoft.Azure.WebJobs.Script.WebHost.Controllers;
-using Microsoft.Azure.WebJobs.Script.WebHost.DependencyInjection;
-using Microsoft.Azure.WebJobs.Script.WebHost.Security;
-using Microsoft.Azure.WebJobs.Script.WebHost.Security.Authentication;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
@@ -45,6 +37,7 @@ namespace Azure.Functions.Cli.Actions.HostActions
         private const int DefaultPort = 7071;
         private const int DefaultTimeout = 20;
         private readonly ISecretsManager _secretsManager;
+        private readonly LogLevel _hostJsonDefaultLogLevel = LogLevel.Information;
 
         public int Port { get; set; }
 
@@ -70,6 +63,13 @@ namespace Azure.Functions.Cli.Actions.HostActions
         public StartHostAction(ISecretsManager secretsManager)
         {
             _secretsManager = secretsManager;
+            try
+            {
+                _hostJsonDefaultLogLevel = Utilities.GetHostJsonDefaultLogLevel(FileSystemHelpers.ReadAllTextFromFile("host.json"));
+            }
+            catch
+            {
+            }
         }
 
         public override ICommandLineParserResult ParseArgs(string[] args)
@@ -171,10 +171,13 @@ namespace Azure.Functions.Cli.Actions.HostActions
                 .ConfigureLogging(loggingBuilder =>
                 {
                     loggingBuilder.ClearProviders();
-                    loggingBuilder.AddDefaultWebJobsFilters();
-                    loggingBuilder.AddProvider(new ColoredConsoleLoggerProvider((cat, level) => level >= LogLevel.Information));
+                    if (_hostJsonDefaultLogLevel != LogLevel.None)
+                    {
+                        loggingBuilder.AddDefaultWebJobsFilters();
+                        loggingBuilder.AddProvider(new ColoredConsoleLoggerProvider((cat, level) => level >= LogLevel.Information));
+                    }
                 })
-                .ConfigureServices((context, services) => services.AddSingleton<IStartup>(new Startup(context, hostOptions, CorsOrigins, CorsCredentials, EnableAuth)))
+                .ConfigureServices((context, services) => services.AddSingleton<IStartup>(new Startup(context, hostOptions, CorsOrigins, CorsCredentials, EnableAuth, _hostJsonDefaultLogLevel)))
                 .Build();
         }
 
@@ -466,123 +469,6 @@ namespace Azure.Functions.Cli.Actions.HostActions
                 ? await SecurityHelpers.GetOrCreateCertificate(CertPath, CertPassword)
                 : null;
             return (new Uri($"{protocol}://0.0.0.0:{Port}"), new Uri($"{protocol}://localhost:{Port}"), cert);
-        }
-
-        public class Startup : IStartup
-        {
-            private readonly WebHostBuilderContext _builderContext;
-            private readonly ScriptApplicationHostOptions _hostOptions;
-            private readonly string[] _corsOrigins;
-            private readonly bool _corsCredentials;
-            private readonly bool _enableAuth;
-
-            public Startup(WebHostBuilderContext builderContext, ScriptApplicationHostOptions hostOptions, string corsOrigins, bool corsCredentials, bool enableAuth)
-            {
-                _builderContext = builderContext;
-                _hostOptions = hostOptions;
-                _enableAuth = enableAuth;
-
-                if (!string.IsNullOrEmpty(corsOrigins))
-                {
-                    _corsOrigins = corsOrigins.Split(',', StringSplitOptions.RemoveEmptyEntries);
-                    _corsCredentials = corsCredentials;
-                }
-            }
-
-            public IServiceProvider ConfigureServices(IServiceCollection services)
-            {
-                if (_corsOrigins != null)
-                {
-                    services.AddCors();
-                }
-
-                if (_enableAuth)
-                {
-                    services.AddWebJobsScriptHostAuthentication();
-                }
-                else
-                {
-                    services.AddAuthentication()
-                        .AddScriptJwtBearer()
-                        .AddScheme<AuthenticationLevelOptions, CliAuthenticationHandler<AuthenticationLevelOptions>>(AuthLevelAuthenticationDefaults.AuthenticationScheme, configureOptions: _ => { })
-                        .AddScheme<ArmAuthenticationOptions, CliAuthenticationHandler<ArmAuthenticationOptions>>(ArmAuthenticationDefaults.AuthenticationScheme, _ => { });
-                }
-
-                services.AddWebJobsScriptHostAuthorization();
-
-                services.AddMvc()
-                    .AddApplicationPart(typeof(HostController).Assembly);
-
-                // workaround for https://github.com/Azure/azure-functions-core-tools/issues/2097
-                SetBundlesEnvironmentVariables();
-
-                services.AddWebJobsScriptHost(_builderContext.Configuration);
-
-                services.Configure<ScriptApplicationHostOptions>(o =>
-                {
-                    o.ScriptPath = _hostOptions.ScriptPath;
-                    o.LogPath = _hostOptions.LogPath;
-                    o.IsSelfHost = _hostOptions.IsSelfHost;
-                    o.SecretsPath = _hostOptions.SecretsPath;
-                });
-
-                services.AddSingleton<IConfigureBuilder<IConfigurationBuilder>>(_ => new ExtensionBundleConfigurationBuilder(_hostOptions));
-                services.AddSingleton<IConfigureBuilder<IConfigurationBuilder>, DisableConsoleConfigurationBuilder>();
-                services.AddSingleton<IConfigureBuilder<ILoggingBuilder>, LoggingBuilder>();
-
-                services.AddSingleton<IDependencyValidator, ThrowingDependencyValidator>();
-
-                return services.BuildServiceProvider();
-            }
-
-            private void SetBundlesEnvironmentVariables()
-            {
-                var bundleId = ExtensionBundleHelper.GetExtensionBundleOptions(_hostOptions).Id;
-                if (!string.IsNullOrEmpty(bundleId))
-                {
-                    Environment.SetEnvironmentVariable("AzureFunctionsJobHost__extensionBundle__downloadPath", ExtensionBundleHelper.GetDownloadPath(bundleId));
-                    Environment.SetEnvironmentVariable("AzureFunctionsJobHost__extensionBundle__ensureLatest", "true");
-                }
-            }
-
-            public void Configure(IApplicationBuilder app)
-            {
-                if (_corsOrigins != null)
-                {
-                    app.UseCors(builder =>
-                    {
-                        var origins = builder.WithOrigins(_corsOrigins)
-                            .AllowAnyHeader()
-                            .AllowAnyMethod();
-                        if (_corsCredentials)
-                        {
-                            origins.AllowCredentials();
-                        }
-                    });
-                }
-
-                IApplicationLifetime applicationLifetime = app.ApplicationServices
-                    .GetRequiredService<IApplicationLifetime>();
-
-                app.UseWebJobsScriptHost(applicationLifetime);
-            }
-
-            private class ThrowingDependencyValidator : DependencyValidator
-            {
-                public override void Validate(IServiceCollection services)
-                {
-                    try
-                    {
-                        base.Validate(services);
-                    }
-                    catch (InvalidHostServicesException ex)
-                    {
-                        // Rethrow this as an InvalidOperationException to bypass the handling
-                        // in the host. This will stop invalid services in the CLI only.
-                        throw new InvalidOperationException("Invalid host services.", ex);
-                    }
-                }
-            }
         }
     }
 }
