+++ src/Azure.Functions.Cli/Diagnostics/LoggingBuilder.cs
@@ -10,9 +10,23 @@ namespace Azure.Functions.Cli.Diagnostics
 {
     internal class LoggingBuilder : IConfigureBuilder<ILoggingBuilder>
     {
+        private LogLevel _hostJsonDefaultLogLevel = LogLevel.Information;
+
+        public LoggingBuilder(LogLevel logLevel)
+        {
+            _hostJsonDefaultLogLevel = logLevel;
+        }
+
         public void Configure(ILoggingBuilder builder)
         {
-            builder.AddProvider(new ColoredConsoleLoggerProvider());
+            if (_hostJsonDefaultLogLevel == LogLevel.None)
+            {
+                builder.AddProvider(new ColoredConsoleLoggerProvider(_hostJsonDefaultLogLevel)).AddFilter((cat, level) => true);
+            }
+            else
+            {
+                builder.AddProvider(new ColoredConsoleLoggerProvider());
+            }
 
             builder.Services.AddSingleton<TelemetryClient>(provider =>
             {
